Move nested component check to googlefonts
