+++ tests/profiles/glyf_test.py
@@ -69,19 +69,3 @@ def test_check_glyf_non_transformed_duplicate_components():
     ttFont['glyf']['quotedbl'].components[1].y = 0
     assert_results_contain(check(ttFont),
                            FAIL, 'found-duplicates')
-
-
-def test_check_glyf_nested_components():
-    """Check glyphs do not have nested components."""
-    check = CheckTester(opentype_profile,
-                        "com.google.fonts/check/glyf_nested_components")
-
-    ttFont = TTFont(TEST_FILE("nunito/Nunito-Regular.ttf"))
-    assert_PASS(check(ttFont))
-
-    # We need to create a nested component. "second" has components, so setting
-    # one of "quotedbl"'s components to "second" should do it.
-    ttFont['glyf']['quotedbl'].components[0].glyphName = "second"
-
-    assert_results_contain(check(ttFont),
-                           FAIL, 'found-nested-components')
